convert to double, finish Exchange

+++ StockMarketSimulator.API/Controllers/WalletController.cs
@@ -23,14 +23,14 @@ namespace StockMarketSimulator.API.Controllers
 
 
         [HttpPost("deposit/{walletId}/{stockSymbol}/{quantity}")]
-        public async Task<IActionResult> Deposit(Guid walletId, string stockSymbol, float quantity)
+        public async Task<IActionResult> Deposit(Guid walletId, string stockSymbol, double quantity)
         {
             await _walletsService.Deposit(walletId, stockSymbol, quantity);
             return Ok();
         }
 
         [HttpPost("withdraw/{walletId}/{stockSymbol}/{quantity}")]
-        public async Task<IActionResult> Withdraw(Guid walletId, string stockSymbol, float quantity)
+        public async Task<IActionResult> Withdraw(Guid walletId, string stockSymbol, double quantity)
         {
             await _walletsService.Withdraw(walletId, stockSymbol, quantity);
             return Ok();
@@ -38,7 +38,7 @@ namespace StockMarketSimulator.API.Controllers
 
 
         [HttpPost("exchange/{walletId}/{baseSymbol}/{quoteSymbol}/{quantity}")]
-        public async Task<IActionResult> Exchange(Guid walletId, string baseSymbol, string quoteSymbol, float quantity)
+        public async Task<IActionResult> Exchange(Guid walletId, string baseSymbol, string quoteSymbol, double quantity)
         {
             await _walletsService.Exchange(walletId, baseSymbol, quoteSymbol, quantity);
             return Ok();
